create light block
