+++ types/block.go
@@ -987,116 +987,6 @@ func CommitFromProto(cp *tmproto.Commit) (*Commit, error) {
 
 //-----------------------------------------------------------------------------
 
-// SignedHeader is a header along with the commits that prove it.
-// It is the basis of the light client.
-type SignedHeader struct {
-	*Header `json:"header"`
-
-	Commit *Commit `json:"commit"`
-}
-
-// ValidateBasic does basic consistency checks and makes sure the header
-// and commit are consistent.
-//
-// NOTE: This does not actually check the cryptographic signatures.  Make sure
-// to use a Verifier to validate the signatures actually provide a
-// significantly strong proof for this header's validity.
-func (sh SignedHeader) ValidateBasic(chainID string) error {
-	if sh.Header == nil {
-		return errors.New("missing header")
-	}
-	if sh.Commit == nil {
-		return errors.New("missing commit")
-	}
-
-	if err := sh.Header.ValidateBasic(); err != nil {
-		return fmt.Errorf("invalid header: %w", err)
-	}
-	if err := sh.Commit.ValidateBasic(); err != nil {
-		return fmt.Errorf("invalid commit: %w", err)
-	}
-
-	if sh.ChainID != chainID {
-		return fmt.Errorf("header belongs to another chain %q, not %q", sh.ChainID, chainID)
-	}
-
-	// Make sure the header is consistent with the commit.
-	if sh.Commit.Height != sh.Height {
-		return fmt.Errorf("header and commit height mismatch: %d vs %d", sh.Height, sh.Commit.Height)
-	}
-	if hhash, chash := sh.Hash(), sh.Commit.BlockID.Hash; !bytes.Equal(hhash, chash) {
-		return fmt.Errorf("commit signs block %X, header is block %X", chash, hhash)
-	}
-
-	return nil
-}
-
-// String returns a string representation of SignedHeader.
-func (sh SignedHeader) String() string {
-	return sh.StringIndented("")
-}
-
-// StringIndented returns an indented string representation of SignedHeader.
-//
-// Header
-// Commit
-func (sh SignedHeader) StringIndented(indent string) string {
-	return fmt.Sprintf(`SignedHeader{
-%s  %v
-%s  %v
-%s}`,
-		indent, sh.Header.StringIndented(indent+"  "),
-		indent, sh.Commit.StringIndented(indent+"  "),
-		indent)
-}
-
-// ToProto converts SignedHeader to protobuf
-func (sh *SignedHeader) ToProto() *tmproto.SignedHeader {
-	if sh == nil {
-		return nil
-	}
-
-	psh := new(tmproto.SignedHeader)
-	if sh.Header != nil {
-		psh.Header = sh.Header.ToProto()
-	}
-	if sh.Commit != nil {
-		psh.Commit = sh.Commit.ToProto()
-	}
-
-	return psh
-}
-
-// FromProto sets a protobuf SignedHeader to the given pointer.
-// It returns an error if the hader or the commit is invalid.
-func SignedHeaderFromProto(shp *tmproto.SignedHeader) (*SignedHeader, error) {
-	if shp == nil {
-		return nil, errors.New("nil SignedHeader")
-	}
-
-	sh := new(SignedHeader)
-
-	if shp.Header != nil {
-		h, err := HeaderFromProto(shp.Header)
-		if err != nil {
-			return nil, err
-		}
-		sh.Header = &h
-	}
-
-	if shp.Commit != nil {
-		c, err := CommitFromProto(shp.Commit)
-		if err != nil {
-			return nil, err
-		}
-		sh.Commit = c
-	}
-
-	return sh, nil
-}
-
-//-----------------------------------------------------------------------------
-
 // Data contains the set of transactions included in the block
 type Data struct {
